backend/requirements.txt
Flask
pymongo
bcrypt
mysql-connector-python
flask-cors